requirements.txt
langgraph
langchain-core
langchain-openai
langchain
akshare
pydantic
python-dotenv
pandas
matplotlib
tqdm
